Added skin changer. The skins files are now located in MTA San Andreas/skins/[Skin Name]. These files are CGUI.is.xml, CGUI.xml, CGUI.lnf.xml and CGUI.png from the existing mta/cgui directory. The default skin is called 'Square', so copying these files to skins/Square will make MTA work. I'll upload a data installer when I work out how to.
Skin can be changed in the settings on the Interface tab (previously called Chat). This now scrolls too, as it's got too big - we need a better solution than that ideally.

You can also change skin by typing current_skin = SomeName into the console.

Changing skin is slow on debug builds (~10s), but almost instant on release builds.

If an invalid skin is chosen, the default skin is loaded. If that can't be loaded, then the user is given an error and MTA quits.

The installer nsi has been updated too.